Develop - sprint 4 - lippujen ja myynnin luominen, katsomienn ja muuttaminen

+++ src/main/java/projekti/demo/model/MyyntiRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MyyntiRepository extends JpaRepository<Myynti, Integer> {
+public interface MyyntiRepository extends JpaRepository<Myynti, Long> {
   
 }
